Divided ROS and V4L code 	- Created the 'camera_driver' module encapsulating the 	  V4L/FFMPEG/OpenCV dedicated code 	- Fixed library overlinking 	- Typo fixes 	- Parameter defaults fixed
